+++ app/components/FAQSection.tsx
