+++ app/components/StatsSection.tsx
