+++ app/components/account/AccountSlice.tsx
@@ -1,6 +1,6 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { produce } from "immer";
-import { RootState } from "../api/store";
+import { RootState } from "../../api/store";
 import {
     fetchAccounts,
     createAccount,
