+++ app/components/car-model/3DCarModel.tsx
