+++ app/components/home/SampleProjects.tsx
