+++ app/team/ic/ICLayout.tsx
@@ -1,4 +1,4 @@
-import { Config } from "../../../../types/ICPageConfig";
+import { Config } from "../../../types/ICPageConfig";
 
 export default function ICLayout({ config }: { config: Config }) {
     return (
